requirements.txt
comtypes
pypdf
PyMuPDF
numpy

reportlab
pillow